chore: transfer event is a v2 feature
